backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "langgraph>=0.2.30",
    "langchain>=0.3.20",
    "langchain-groq>=0.2.4",
    "langchain-huggingface>=0.1.2",
    "langchain-qdrant>=0.1.4",
    "langchain-community>=0.3.20",
    "qdrant-client>=1.10.0",
    "tavily-python>=0.5.0",
    "beautifulsoup4>=4.12.3",
    "pydantic>=2.7.0",
    "python-decouple>=3.8",
    "langchain-text-splitters>=1.1.1",
    "openrouteservice>=0.4.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]
